Fix more sprintf formats (or args)
git-svn-id: https://svn.isis.vanderbilt.edu/GMESRC/trunk@2811 b35830b6-564f-0410-ac5f-b387e50cb686

+++ GME/XmlBackEnd/CoreXmlFile.cpp
@@ -5257,7 +5257,7 @@ bool CCoreXmlFile::makeSureFileExistsInVerSys( const std::string& p_fname, const
 	}
 	catch( hresult_exception& e)
 	{
-		char buff[200]; sprintf( buff, "Could not get \"%s\" file from source control. Exception code: 0x%x", fulllocalfname.c_str(), e.hr);
+		char buff[200]; sprintf( buff, "Could not get \"%s\" file from source control. Exception code: 0x%x", fulllocalfname.c_str(), (unsigned int)e.hr);
 		sendMsg( buff, MSG_ERROR);
 		return false;
 	}
